+++ src/extension.ts
@@ -3,17 +3,19 @@ import { HUDPanelProvider } from './hudPanel';
 import { StatusBarManager } from './statusBar';
 import { DataProvider } from './dataProvider';
 import { ConfigManager } from './configManager';
+import { HistoryStore } from './historyStore';
 
 let tickTimer: ReturnType<typeof setInterval> | undefined;
 
 export function activate(context: vscode.ExtensionContext): void {
   // ---- Init components ----
   const configManager = new ConfigManager();
-  const dataProvider = new DataProvider(configManager);
+  const historyStore = new HistoryStore(context);
+  const dataProvider = new DataProvider(configManager, historyStore);
   const statusBar = new StatusBarManager();
 
   // ---- Register webview provider ----
-  const provider = new HUDPanelProvider(context.extensionUri, configManager);
+  const provider = new HUDPanelProvider(context.extensionUri, configManager, context.globalState);
 
   context.subscriptions.push(
     vscode.window.registerWebviewViewProvider(HUDPanelProvider.viewType, provider, {